refactor: refactor authorizationUrl tests into its own file
